Revert file renames to manage diff
