+++ packages/malloy/src/lang/test/syntax-errors.spec.ts
